Add files via upload
The BurgerWorld website is a vibrant online portal created to celebrate the art and joy of hamburger cuisine. Utilizing the foundational web technologies HTML and CSS, this website offers a visually appealing and user-friendly platform where hamburger enthusiasts can explore a variety of burger recipes, cooking tips, and the latest burger trends.